Extracted signature verification to new module
